don't rewrite user-dirs.dirs on every login

+++ base/.profile
@@ -14,8 +14,9 @@ RC="$HOME/.bashrc"
 
 
 # Fix XDG directories to map to correct $HOME
-rm $XDG_CONFIG_HOME/user-dirs.dirs
-sed "s|\$HOME|${HOME}|g" $XDG_CONFIG_HOME/user-dirs.tmp > $XDG_CONFIG_HOME/user-dirs.dirs
+if test ! -f $XDG_CONFIG_HOME/user-dirs.dirs; then
+  sed "s|\$HOME|${HOME}|g" $XDG_CONFIG_HOME/user-dirs.tmp > $XDG_CONFIG_HOME/user-dirs.dirs
+fi
 
 
 # Start X at login on first TTY provided we're not SSH'd in
